Validation msg added not non existing user

+++ src/modules/user/user.service.ts
@@ -2,6 +2,9 @@ import { IUser } from "./user.interface";
 import { User } from "./user.model";
 
 const createUserIntoDB = async (userData: IUser) => {
+  if (await User.findOne({ userId: userData.userId })) {
+    throw new Error("User Already Exits");
+  }
   const result = await User.create(userData);
   return result;
 };
@@ -12,12 +15,20 @@ const getAllUsersFromDB = async () => {
 };
 
 const getSingleUserFromDB = async (id: number) => {
-  const result = await User.findOne({ userId: id }).select("-password");
+  if (!(await User.isUserExists(id))) {
+    throw new Error("User does not exists");
+  }
+  const result = await User.findOne({ userId: id }).select(
+    "-password -orders -fullName._id -__v"
+  );
 
   return result;
 };
 
 const updateUserIntoDB = async (id: number, userData: IUser) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("User does not exists");
+  }
   const result = await User.updateOne({ userId: id }, userData, {
     new: true,
     runValidators: true,
@@ -26,7 +37,11 @@ const updateUserIntoDB = async (id: number, userData: IUser) => {
 };
 
 const deleteUser = async (id: number) => {
+  if (!(await User.isUserExists(id))) {
+    throw new Error("User does not exists");
+  }
   const result = await User.deleteOne({ userId: id });
+  console.log(result);
   return result;
 };
